Sorta final project

+++ COP3503.cpp
@@ -32,7 +32,7 @@ double COP3503::pointSummer(double grades[], int arraySize) {
 }
 double COP3503::examPointSummer(array<double, 20> examGrades) {
 	double total = 0;
-	for (int i = 0; i < examGrades.size(); i++) {
+	for (size_t i = 0; i < examGrades.size(); i++) {
 		//this check makes it so only grades that have been entered are added to the total
 		if (examGrades[i] != -1) {
 			total += examGrades[i];
