+++ MAC2311.cpp
@@ -77,7 +77,7 @@ double MAC2311::pointSummer(double grades[], int arraySize) {
 //this is exactly like pointSummer except that it takes in an array<T, V> type
 double MAC2311::examPointSummer(array<double, 20> examGrades) {
 	double total = 0;
-	for (int i = 0; i < examGrades.size(); i++) {
+	for (size_t i = 0; i < examGrades.size(); i++) {
 		//this check makes it so only grades that have been entered are added to the total
 		if (examGrades[i] != -1) {
 			total += examGrades[i];
